Check number of games in scraped dataset

+++ R/save_pbp.R
@@ -6,17 +6,24 @@ save_pbp <- function(season) {
 
   pbp <- nflfastR::build_nflfastR_pbp(ids, pp = TRUE)
 
-  # rds
-  saveRDS(pbp, glue::glue('data/play_by_play_{season}.rds'))
-  # csv.gz
-  readr::write_csv(pbp, glue::glue('data/play_by_play_{season}.csv.gz'))
-  # .parquet
-  arrow::write_parquet(pbp, glue::glue('data/play_by_play_{season}.parquet'))
-  # .zip
-  readr::write_csv(pbp, glue::glue("data/play_by_play_{season}.csv"))
-  utils::zip(glue::glue("data/play_by_play_{season}.zip"), c(glue::glue("data/play_by_play_{season}.csv")))
-  file.remove(glue::glue("data/play_by_play_{season}.csv"))
-  closeAllConnections()
-  usethis::ui_done("Saved {season} pbp data.")
+  if(length(unique(pbp$game_id)) == length(ids)){
+    # rds
+    saveRDS(pbp, glue::glue('data/play_by_play_{season}.rds'))
+    # csv.gz
+    readr::write_csv(pbp, glue::glue('data/play_by_play_{season}.csv.gz'))
+    # .parquet
+    arrow::write_parquet(pbp, glue::glue('data/play_by_play_{season}.parquet'))
+    # .zip
+    readr::write_csv(pbp, glue::glue("data/play_by_play_{season}.csv"))
+    utils::zip(glue::glue("data/play_by_play_{season}.zip"), c(glue::glue("data/play_by_play_{season}.csv")))
+    file.remove(glue::glue("data/play_by_play_{season}.csv"))
+    closeAllConnections()
+    usethis::ui_done("Saved {season} pbp data.")
+  } else {
+    usethis::ui_warn(c(
+      "The number of finished games is not equal to the number of games in the loaded data.",
+      "Will not push this corrupted dataset to the repo."
+      ))
+  }
   return(invisible())
 }
